+++ clang-tools-extra/clang-tidy/daedalean/DaedaleanTidyModule.cpp
@@ -17,6 +17,7 @@
 #include "FloatingPointComparisonCheck.h"
 #include "IncludeOrderCheck.h"
 #include "LambdaImplicitCaptureCheck.h"
+#include "PreprocessingDirectivesCheck.h"
 #include "ProtectedMustNotBeUsedCheck.h"
 #include "StructsAndClassesCheck.h"
 #include "SwitchStatementCheck.h"
@@ -50,6 +51,8 @@ public:
         "daedalean-include-order");
     CheckFactories.registerCheck<LambdaImplicitCaptureCheck>(
         "daedalean-lambda-implicit-capture");
+    CheckFactories.registerCheck<PreprocessingDirectivesCheck>(
+        "daedalean-preprocessing-directives");
     CheckFactories.registerCheck<ProtectedMustNotBeUsedCheck>(
         "daedalean-protected-must-not-be-used");
     CheckFactories.registerCheck<StructsAndClassesCheck>(